requirements.txt
yt-dlp>=2025.4.30
moviepy>=1.0.3
opencv-python>=4.8.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
